fix: refactor domain-canonical and domain-inputs into their own files
