+++ apps/ensapi/src/omnigraph-api/lib/find-domains/domain-cursor.ts
@@ -2,7 +2,7 @@ import type { DomainId } from "enssdk";
 
 import { cursors } from "@/omnigraph-api/lib/cursors";
 import type { DomainOrderValue } from "@/omnigraph-api/lib/find-domains/types";
-import type { DomainsOrderBy } from "@/omnigraph-api/schema/domain";
+import type { DomainsOrderBy } from "@/omnigraph-api/schema/domain-inputs";
 import type { OrderDirection } from "@/omnigraph-api/schema/order-direction";
 
 /**
